+++ crates/optimism/src/evm.rs
@@ -4,7 +4,7 @@ use revm::{
     context_interface::ContextTr,
     handler::{
         instructions::{EthInstructions, InstructionProvider},
-        EvmTr,
+        EvmTr, PrecompileProvider,
     },
     inspector::{InspectorEvmTr, JournalExt},
     interpreter::{interpreter::EthInterpreter, Interpreter, InterpreterAction, InterpreterTypes},
@@ -32,6 +32,7 @@ where
         Context = CTX,
         InterpreterTypes: InterpreterTypes<Output = InterpreterAction>,
     >,
+    P: PrecompileProvider<CTX>,
     INSP: Inspector<CTX, I::InterpreterTypes>,
 {
     type Inspector = INSP;
@@ -62,6 +63,7 @@ where
         Context = CTX,
         InterpreterTypes: InterpreterTypes<Output = InterpreterAction>,
     >,
+    P: PrecompileProvider<CTX>,
 {
     type Context = CTX;
     type Instructions = I;
@@ -283,7 +285,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_g1_add_out_of_gas() {
         let ctx = Context::op()
             .modify_tx_chained(|tx| {
@@ -311,7 +312,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_g1_add_input_wrong_size() {
         let ctx = Context::op()
             .modify_tx_chained(|tx| {
@@ -370,7 +370,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_g1_msm_input_wrong_size() {
         let ctx = g1_msm_test_tx().modify_tx_chained(|tx| tx.base.data = tx.base.data.slice(1..));
 
@@ -388,7 +387,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_g1_msm_out_of_gas() {
         let ctx = g1_msm_test_tx().modify_tx_chained(|tx| tx.base.gas_limit -= 1);
 
@@ -406,7 +404,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_g1_msm_wrong_input_layout() {
         let ctx = g1_msm_test_tx();
 
@@ -414,6 +411,60 @@ mod tests {
         let output = evm.replay().unwrap();
 
         // assert fails post gas check, because input is wrong layout
+        assert!(matches!(
+            output.result,
+            ExecutionResult::Halt {
+                reason: OpHaltReason::Base(HaltReason::PrecompileError),
+                ..
+            }
+        ));
+    }
+
+    #[test]
+    fn test_halted_tx_call_bls12_381_g2_add_out_of_gas() {
+        let ctx = Context::op()
+            .modify_tx_chained(|tx| {
+                tx.base.kind = TxKind::Call(bls12_381_const::G2_ADD_ADDRESS);
+                tx.base.gas_limit = 21_000 + bls12_381_const::G2_ADD_BASE_GAS_FEE - 1;
+            })
+            .modify_chain_chained(|l1_block| {
+                l1_block.operator_fee_constant = Some(U256::ZERO);
+                l1_block.operator_fee_scalar = Some(U256::ZERO)
+            })
+            .modify_cfg_chained(|cfg| cfg.spec = OpSpecId::ISTHMUS);
+
+        let mut evm = ctx.build_op();
+
+        let output = evm.replay().unwrap();
+
+        // assert out of gas
+        assert!(matches!(
+            output.result,
+            ExecutionResult::Halt {
+                reason: OpHaltReason::Base(HaltReason::OutOfGas(OutOfGasError::Precompile)),
+                ..
+            }
+        ));
+    }
+
+    #[test]
+    fn test_halted_tx_call_bls12_381_g2_add_input_wrong_size() {
+        let ctx = Context::op()
+            .modify_tx_chained(|tx| {
+                tx.base.kind = TxKind::Call(bls12_381_const::G2_ADD_ADDRESS);
+                tx.base.gas_limit = 21_000 + bls12_381_const::G2_ADD_BASE_GAS_FEE;
+            })
+            .modify_chain_chained(|l1_block| {
+                l1_block.operator_fee_constant = Some(U256::ZERO);
+                l1_block.operator_fee_scalar = Some(U256::ZERO)
+            })
+            .modify_cfg_chained(|cfg| cfg.spec = OpSpecId::ISTHMUS);
+
+        let mut evm = ctx.build_op();
+
+        let output = evm.replay().unwrap();
+
+        // assert fails post gas check, because input is wrong size
         assert!(matches!(
             output.result,
             ExecutionResult::Halt {
@@ -456,7 +507,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_g2_msm_input_wrong_size() {
         let ctx = g2_msm_test_tx().modify_tx_chained(|tx| tx.base.data = tx.base.data.slice(1..));
 
@@ -474,7 +524,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_g2_msm_out_of_gas() {
         let ctx = g2_msm_test_tx().modify_tx_chained(|tx| tx.base.gas_limit -= 1);
 
@@ -492,7 +541,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_g2_msm_wrong_input_layout() {
         let ctx = g2_msm_test_tx();
 
@@ -540,7 +588,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_pairing_input_wrong_size() {
         let ctx = bl12_381_pairing_test_tx()
             .modify_tx_chained(|tx| tx.base.data = tx.base.data.slice(1..));
@@ -559,7 +606,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_pairing_out_of_gas() {
         let ctx = bl12_381_pairing_test_tx().modify_tx_chained(|tx| tx.base.gas_limit -= 1);
 
@@ -577,7 +623,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_tx_call_bls12_381_pairing_wrong_input_layout() {
         let ctx = bl12_381_pairing_test_tx();
 
@@ -622,7 +667,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_map_fp_to_g1_out_of_gas() {
         let ctx = fp_to_g1_test_tx().modify_tx_chained(|tx| tx.base.gas_limit -= 1);
 
@@ -640,7 +684,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_map_fp_to_g1_input_wrong_size() {
         let ctx = fp_to_g1_test_tx().modify_tx_chained(|tx| tx.base.data = tx.base.data.slice(1..));
 
@@ -685,7 +728,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_map_fp2_to_g2_out_of_gas() {
         let ctx = fp2_to_g2_test_tx().modify_tx_chained(|tx| tx.base.gas_limit -= 1);
 
@@ -703,7 +745,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_map_fp2_to_g2_input_wrong_size() {
         let ctx =
             fp2_to_g2_test_tx().modify_tx_chained(|tx| tx.base.data = tx.base.data.slice(1..));